agents/orchestrator/requirements.txt
strands-agents>=1.22.0
bedrock-agentcore>=0.1.0
fastapi>=0.115.0
uvicorn>=0.38.0
boto3>=1.42.0
httpx>=0.28.0
mcp>=1.25.0
pydantic>=2.0.0
aws-opentelemetry-distro>=0.10.1
